Refactor: Decompose EnemyShipFormation into MVC pattern
